move ObjectData + Environment
