+++ source/engine/graphics/shaders/CMakeLists.txt
@@ -1,7 +1,6 @@
 target_sources(${NC_ENGINE_LIB}
     PRIVATE
         CubeMapShaderResource.cpp
-        Environment.cpp
         EnvironmentDataShaderResource.cpp
         ObjectDataShaderResource.cpp
         PointLight.cpp
