+++ source/engine/graphics/shaders/ObjectDataShaderResource.h
@@ -1,25 +1,25 @@
 #pragma once
 
-#include "DirectXMath.h"
 #include "graphics/buffers/WriteableBuffer.h"
 #include "graphics/shaders/ShaderDescriptorSets.h"
 #include "graphics/shaders/ShaderResourceService.h"
+#include "graphics/shader_resource/ObjectData.h"
 
 namespace nc::graphics
 {
-    /** @todo We probably don't want the viewProjection in this struct. We only need
-     *  1 copy, not 1 per renderer. */
-    struct ObjectData
-    {
-        DirectX::XMMATRIX model;
-        DirectX::XMMATRIX modelView;
-        DirectX::XMMATRIX viewProjection;
+    // /** @todo We probably don't want the viewProjection in this struct. We only need
+    //  *  1 copy, not 1 per renderer. */
+    // struct ObjectData
+    // {
+    //     DirectX::XMMATRIX model;
+    //     DirectX::XMMATRIX modelView;
+    //     DirectX::XMMATRIX viewProjection;
 
-        uint32_t baseColorIndex; // Todo: Make this more generic for materials
-        uint32_t normalIndex;  // Todo: Make this more generic for materials
-        uint32_t roughnessIndex;  // Todo: Make this more generic for materials
-        uint32_t metallicIndex;  // Todo: Make this more generic for materials
-    };
+    //     uint32_t baseColorIndex; // Todo: Make this more generic for materials
+    //     uint32_t normalIndex;  // Todo: Make this more generic for materials
+    //     uint32_t roughnessIndex;  // Todo: Make this more generic for materials
+    //     uint32_t metallicIndex;  // Todo: Make this more generic for materials
+    // };
 
     class ObjectDataShaderResource : public IShaderResource<ObjectData>
     {
