+++ source/engine/graphics/techniques/ShadowMappingTechnique.cpp
@@ -1,12 +1,14 @@
 #include "ShadowMappingTechnique.h"
 #include "config/Config.h"
-#include "graphics/MeshRenderer.h"
 #include "graphics/GpuOptions.h"
-#include "graphics/meshes/VertexDescriptions.h"
 #include "graphics/Initializers.h"
+#include "graphics/MeshRenderer.h"
+#include "graphics/PerFrameRenderState.h"
+#include "graphics/meshes/VertexDescriptions.h"
 #include "graphics/shaders/ShaderDescriptorSets.h"
 #include "graphics/shaders/ShaderResources.h"
 #include "graphics/shaders/ShaderUtilities.h"
+
 #include "optick/optick.h"
 
 namespace
